backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - Supabase
supabase==2.3.4
postgrest==0.13.2

# AI & ML - Using Gemini (FREE!)
google-generativeai==0.3.2
langchain-google-genai==0.0.5
langchain==0.1.5
prophet==1.1.5
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# ZYND AI Agent
zyndai-agent==0.1.0

# Weather & Geo APIs (removed httpx to fix conflict)
requests==2.31.0
geopy==2.4.1

# WebSockets & Real-time
python-socketio==5.11.0
websockets==12.0

# Background Tasks (Optional - uncomment if needed)
# celery==5.3.6
# redis==5.0.1

# Auth & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.1.1
